requirements.txt
absl-py==1.2.0
anyio==3.6.2
appdirs==1.4.4
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1640817743617/work
argon2-cffi-bindings @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi-bindings_1649500366499/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1660605382950/work
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1659291887007/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1649463573192/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1656355450470/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.2.0
certifi==2022.9.24
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1656782892388/work
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1660619107463/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dill==0.3.5.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
etils==0.8.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1665301981797/work
fastapi==0.88.0
fastjsonschema @ file:///home/conda/feedstock_root/build_artifacts/python-fastjsonschema_1663619548554/work/dist
Flask==2.2.2
flatbuffers==22.9.24
flit_core @ file:///home/conda/feedstock_root/build_artifacts/flit-core_1645629044586/work/source/flit_core
fonttools==4.37.4
gast==0.4.0
gevent==22.10.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1653138941901/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964045648/work
idna==2.8
importlib-metadata==5.0.0
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1665204935269/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1664214831245/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1662481590184/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1662482321563/work
itsdangerous==2.1.2
jax==0.3.25
jaxlib==0.3.25
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1659959867326/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
joblib==1.2.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema-meta_1662736858348/work
jupyter @ file:///Users/runner/miniforge3/conda-bld/jupyter_1637233473259/work
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1655961255101/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1661522530937/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1658332720484/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1649936611996/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1662157840858/work
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1648737598016/work
matplotlib==3.6.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1635844817800/work
multitasking==0.0.11
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1646999386773/work
nbconvert @ file:///Users/runner/miniforge3/conda-bld/nbconvert_1648465200908/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1664204096384/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1654636967533/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1660672958449/work
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.5.0
pandas-datareader==0.10.0
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.2.0
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1633981968097/work
portpicker==1.2.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1649447152425/work
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1662356191042/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.10.2
pydot==1.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1649013408545/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1664798238822/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1663830604151/work
qtconsole==5.3.2
QtPy==2.2.1
requests==2.21.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
simplegeneric==0.8.1
six==1.12.0
sniffio==1.3.0
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1658207591808/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1664126450622/work
starlette==0.22.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.7.0
tensorflow-decision-forests==1.0.1
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-macos==2.10.0
tensorflow-metadata==1.10.0
tensorflow-metal==0.6.0
tensorflowjs==4.1.0
termcolor==2.0.1
terminado==0.13.3
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1645693042223/work
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1663005918942/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3==1.24.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1662157836868/work
wrapt==1.14.1
wurlitzer==3.0.3
yfinance==0.1.77
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1665306627898/work
zope.event==4.5.0
zope.interface==5.5.2
